requirements.txt
Django==3.1.6
django-crispy-forms
python-dotenv
django-autoslug
pillow